[#142] refactor: 댓글 atom 초기값 분리
