+++ src/hooks/atoms.ts
@@ -1,0 +1,6 @@
+import { atom } from 'jotai';
+
+// 댓글 카운트
+const commentCountAtom = atom(0);
+
+export { commentCountAtom };
